sanitize-commit: Recognize "QUIP" as a footer
Having no hyphen and not being one of the known hyphen-less footer
names, "QUIP" was not recognized as a footer.

Fix by including it in the list of known footer names. Relaxing the
regexp to not require a hyphen would probably cause too many
false-positives.

See various uses of QUIP: 19 and QUIP: 23 as footers in merged commits.

Fixes: QTQAINFRA-7033
Change-Id: I57ec60c0a8056b196fba3ce70ffab45cf2fffe2f
Reviewed-by: Daniel Smith <daniel.smith@qt.io>