util/normalize: parse at the 8-bit level
Like so much Qt code overusing UTF-16, the old code converted each
line into a QString, and then, to feed into normalizedSignature(),
back into 8-bit, and then back into UTF-16 and then, when modifying,
back to 8-bit.

Just do everything in 8-bit. Yes, this means setFöö() will likely not
work anymore when source code is encoded in UTF-8, but MOC can't store
that, anyway, AFAIK, so we're not losing anything. If it becomes a
problem, it's easy to fix in isValidIdentifier().

Change-Id: Ibf03032ed892c968db3db2f28c188455317d3377
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>