sanitize-commit: don't make unknown cherry-pick SHA-1's fatal
The SHA-1 might just be from the tqtc- version of the repo and so this
bot complaint should be overridable by reviewers. So don't issue a -2
in this case.

We had at least one cherry-pick from lts-6.5 to dev where this is
blocking.

As a drive-by, on reviewer request, also change the warning message
itself.

Change-Id: Ifc7c4c5129bf9ad921bb6aba737b1a7076476d2c
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>